:test_tube: Refactors functions to test helpers
